requirements.txt
# Core UI & Agent Framework
streamlit==1.28.1
openai==1.3.5
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.2.2
pandas==2.1.3
numpy==1.25.2
pydantic==2.5.2
tenacity==8.2.3
aiohttp==3.9.1

# Google Calendar Integration
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Multi-Agent System
asyncio-mqtt==0.11.1
python-dateutil==2.8.2

# Database & Vector Store
psycopg2-binary==2.9.9
pgvector==0.2.4
redis==5.0.1
chromadb==0.4.18

# Observability & Monitoring
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
sentry-sdk==1.39.1
